Made separate file for company locations model.
