+++ meteor-app/imports/api/models/index.js
@@ -3,6 +3,7 @@ export * from "./comment.js";
 export * from "./misc.js";
 export * from "./company.js";
 export * from "./job-ad.js";
+export * from "./location.js";
 export * from "./review.js";
 export * from "./salary.js";
 export * from "./user.js";
