+++ meteor-app/imports/api/models/misc.js
@@ -2,12 +2,6 @@
 
 export type ID = string;
 
-export type Location = {
-	city: string,
-	address: string,
-	industrialHub: ?string,
-};
-
 export type StarRatings = {
 	healthAndSafety: number,
 	managerRelationship: number,
